sync-process-docs/pyproject.toml
[project]
name = "sync-process-docs-runtime"
version = "0.1.0"
description = "Sync document processing workflow using Google Gemini to extract information from documents one at a time"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exospherehost>=0.0.2",
    "python-dotenv>=1.1.1",
    "google-genai>=0.3.0",
    "requests>=2.28.0",
    "pydantic>=1.10.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pandas>=2.0.0",
    "jsonschema>=4.0.0",
    "motor>=3.7.1",
    "pymongo>=4.0.0",
    "pypdf2>=3.0.1",
    "google-genai>=1.33.0",
]
